server-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.67.0,<3.0.0
python-dotenv==1.0.0
pydantic>=2.10,<3.0.0
python-multipart==0.0.6
pydantic-ai-slim[openai]==0.0.49
pydantic-graph==0.0.49
asyncpg==0.29.0

